refactor: Introduce new module and rename
