+++ src-docs/reactive.runner_manager.md
@@ -3,53 +3,54 @@
 <a href="../src/github_runner_manager/reactive/runner_manager.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 # <kbd>module</kbd> `reactive.runner_manager`
-Module for managing reactive runners. 
+Module for reconciling amount of runner and reactive runner processes. 
 
-**Global Variables**
----------------
-- **PYTHON_BIN**
-- **REACTIVE_RUNNER_SCRIPT_MODULE**
-- **REACTIVE_RUNNER_CMD_LINE_PREFIX**
-- **PID_CMD_COLUMN_WIDTH**
-- **PIDS_COMMAND_LINE**
-- **UBUNTU_USER**
-- **RUNNER_CONFIG_ENV_VAR**
 
 ---
 
-<a href="../src/github_runner_manager/reactive/runner_manager.py#L40"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+<a href="../src/github_runner_manager/reactive/runner_manager.py#L14"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 ## <kbd>function</kbd> `reconcile`
 
 ```python
-reconcile(quantity: int, runner_config: RunnerConfig) → int
+reconcile(
+    quantity: int,
+    runner_manager: RunnerManager,
+    runner_config: RunnerConfig
+) → int
 ```
 
-Spawn a runner reactively. 
+Reconcile runners reactively. 
+
+The reconciliation attempts to make the following equation true:  quantity_of_current_runners + reactive_processes_consuming_jobs == quantity. 
+
+A few examples: 
+
+1. If there are 5 runners and 5 reactive processes and the quantity is 10,  no action is taken. 2. If there are 5 runners and 5 reactive processes and the quantity is 15,  5 reactive processes are created. 3. If there are 5 runners and 5 reactive processes and quantity is 7,  3 reactive processes are killed. 4. If there are 5 runners and 5 reactive processes and quantity is 5,  all reactive processes are killed. 5. If there are 5 runners and 5 reactive processes and quantity is 4,  1 runner is killed and all reactive processes are killed. 
+
+
+
+So if the quantity is equal to the sum of the current runners and reactive processes, no action is taken, 
+
+If the quantity is greater than the sum of the current runners and reactive processes, additional reactive processes are created. 
+
+If the quantity is greater than or equal to the quantity of the current runners, but less than the sum of the current runners and reactive processes, additional reactive processes will be killed. 
+
+If the quantity is less than the sum of the current runners, additional runners are killed and all reactive processes are killed. 
+
+In addition to this behaviour, reconciliation also checks the queue at the start and removes all idle runners if the queue is empty, to ensure that no idle runners are left behind if there are no new jobs. 
 
 
 
 **Args:**
  
- - <b>`quantity`</b>:  The number of runners to spawn. 
- - <b>`runner_config`</b>:  The reactive runner configuration. 
-
-Raises a ReactiveRunnerError if the runner fails to spawn. 
+ - <b>`quantity`</b>:  Number of intended amount of runners + reactive processes. 
+ - <b>`runner_manager`</b>:  The runner manager to interact with current running runners. 
+ - <b>`runner_config`</b>:  The reactive runner config. 
 
 
 
 **Returns:**
- The number of reactive runner processes spawned. 
-
-
----
-
-<a href="../src/github_runner_manager/reactive/runner_manager.py#L36"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
-
-## <kbd>class</kbd> `ReactiveRunnerError`
-Raised when a reactive runner error occurs. 
-
-
-
+ The number of reactive processes created. If negative, its absolute value is equal to the number of processes killed.
