+++ src/github_runner_manager/manager/runner_scaler.py
@@ -126,7 +126,11 @@ class RunnerScaler:
 
         if self._reactive_config is not None:
             logger.info("Reactive configuration detected, going into experimental reactive mode.")
-            return self._reconcile_reactive(quantity)
+            return reactive_runner_manager.reconcile(
+                quantity=quantity,
+                runner_manager=self._manager,
+                runner_config=self._reactive_config,
+            )
 
         metric_stats = {}
         start_timestamp = time.time()
@@ -248,54 +252,3 @@ class RunnerScaler:
             )
         except IssueMetricEventError:
             logger.exception("Failed to issue Reconciliation metric")
-
-    def _reconcile_reactive(self, quantity: int) -> int:
-        """Reconcile runners reactively.
-
-        The reconciliation attempts to make the following equation true:
-            quantity_of_current_runners + reactive_processes_consuming_jobs == quantity.
-
-        A few examples:
-
-        1. If there are 5 runners and 5 reactive processes and the quantity is 10,
-            no action is taken.
-        2. If there are 5 runners and 5 reactive processes and the quantity is 15,
-            5 reactive processes are created.
-        3. If there are 5 runners and 5 reactive processes and quantity is 7,
-            3 reactive processes are killed.
-        4. If there are 5 runners and 5 reactive processes and quantity is 5,
-            all reactive processes are killed.
-        5. If there are 5 runners and 5 reactive processes and quantity is 4,
-            1 runner is killed and all reactive processes are killed.
-
-
-        So if the quantity is equal to the sum of the current runners and reactive processes,
-        no action is taken,
-
-        If the quantity is greater than the sum of the current
-        runners and reactive processes, additional reactive processes are created.
-
-        If the quantity is greater than or equal to the quantity of the current runners,
-        but less than the sum of the current runners and reactive processes,
-        additional reactive processes will be killed.
-
-        If the quantity is less than the sum of the current runners,
-        additional runners are killed and all reactive processes are killed.
-
-        In addition to this behaviour, reconciliation also checks the queue at the start and
-        removes all idle runners if the queue is empty, to ensure that
-        no idle runners are left behind if there are no new jobs.
-
-        Args:
-            quantity: Number of intended amount of runners + reactive processes.
-
-        Returns:
-            The number of reactive processes created. If negative, its absolute value is equal
-            to the number of processes killed.
-        """
-        logger.info("Reactive mode is experimental and not yet fully implemented.")
-        self._manager.cleanup()
-        return reactive_runner_manager.reconcile(
-            quantity=quantity,
-            runner_config=self._reactive_config,
-        )
